PoC with flag

+++ silo-core/test/foundry/interestRateModel/InterestRateModelV2Impl.sol
@@ -5,14 +5,14 @@ import {InterestRateModelV2} from "silo-core/contracts/interestRateModel/Interes
 
 contract InterestRateModelV2Impl is InterestRateModelV2 {
     function mockSetup(address _silo, int256 _ri, int256 _Tcrit) external {
-        if (_Tcrit > type(int128).max) revert("[InterestRateModelV2Impl] _Tcrit overflow");
-        if (_Tcrit < type(int128).min) revert("[InterestRateModelV2Impl] _Tcrit underflow");
+        if (_Tcrit > type(int112).max) revert("[InterestRateModelV2Impl] _Tcrit overflow");
+        if (_Tcrit < type(int112).min) revert("[InterestRateModelV2Impl] _Tcrit underflow");
 
-        if (_ri > type(int128).max) revert("[InterestRateModelV2Impl] _ri overflow");
-        if (_ri < type(int128).min) revert("[InterestRateModelV2Impl] _ri underflow");
+        if (_ri > type(int112).max) revert("[InterestRateModelV2Impl] _ri overflow");
+        if (_ri < type(int112).min) revert("[InterestRateModelV2Impl] _ri underflow");
 
-        getSetup[_silo].Tcrit = int128(_Tcrit);
-        getSetup[_silo].ri = int128(_ri);
+        getSetup[_silo].Tcrit = int112(_Tcrit);
+        getSetup[_silo].ri = int112(_ri);
     }
 
     function calculateRComp(
